wip: pull import resolution into separate pass
